requirements.txt
langchain
langchain-openai
newspaper3k
python-dotenv
